requirements.txt
# Django
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0

 # Photo & Image
 Pillow==10.1.0

 # Http request
 requests==2.31.0